+++ src/my_shell.c
@@ -39,7 +39,7 @@ static void exec(char *str, shell_t *shell)
     free(str);
     str = NULL;
     vec_t *pid = redirection(replace_alias(vec, shell), shell);
-    for (int i = 0; pid != NULL && i < pid->element; i++) {
+    for (int i = 0; pid != NULL && i < (int) pid->element; i++) {
         int s = 0;
         waitpid(*(int *)(pid->content[i]), &s, 0);
         (WIFSIGNALED(s))?print_err(s, shell):
